rename DatabaseObservationDAO to ObservationDAO and remove interface; rename DatabaseReadAccessDAO to ReadAccessDAO
