+++ caom2harvester/build.gradle
@@ -27,7 +27,7 @@ dependencies {
     compile 'org.opencadc:cadc-util:[1.0,)'
     compile 'org.opencadc:caom2:[2.3.4,)'
     compile 'org.opencadc:caom2-compute:[2.3.3,)'
-    compile 'org.opencadc:caom2persistence:[2.3.5,)'
+    compile 'org.opencadc:caom2persistence:[2.3.6,)'
     compile 'org.opencadc:cadc-util:[1.0.14,)'
     compile 'org.opencadc:caom2-repo:[1.0,)'
     compile 'org.opencadc:caom2-persist:[2.3.1,)'
