+++ caom2persistence/build.gradle
@@ -14,7 +14,7 @@ sourceCompatibility = 1.7
 
 group = 'org.opencadc'
 
-version = '2.3.5'
+version = '2.3.6'
 
 dependencies {
     compile 'log4j:log4j:1.2.+'
