+++ caom2persistence/src/test/java/ca/nrc/cadc/caom2/persistence/AbstractDatabaseReadAccessDAOTest.java
@@ -103,8 +103,8 @@ public abstract class AbstractDatabaseReadAccessDAOTest
 
     boolean deletionTrack;
     boolean useLongForUUID;
-    DatabaseReadAccessDAO dao;
-    DatabaseObservationDAO obsDAO;
+    ReadAccessDAO dao;
+    ObservationDAO obsDAO;
     DatabaseTransactionManager txnManager;
 
     protected Class[] entityClasses;
@@ -121,11 +121,11 @@ public abstract class AbstractDatabaseReadAccessDAOTest
             config.put("database", database);
             config.put("schema", schema);
             config.put(SQLGenerator.class.getName(), genClass);
-            this.dao = new DatabaseReadAccessDAO();
+            this.dao = new ReadAccessDAO();
             dao.setConfig(config);
             this.txnManager = new DatabaseTransactionManager(dao.getDataSource());
             
-            this.obsDAO = new DatabaseObservationDAO();
+            this.obsDAO = new ObservationDAO();
             obsDAO.setConfig(config);
         }
         catch(Exception ex)
